[Q20-3] - moved classes to appropriate packages and removed unnecessary ones
