--- quadriga/src/main/java/edu/asu/diging/quadriga/config/web/ApiAuthenticationDetailsSource.java
@@ -1,18 +1,0 @@
-package edu.asu.diging.quadriga.config.web;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.springframework.security.authentication.AuthenticationDetailsSource;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.web.authentication.www.BasicAuthenticationConverter;
-
-public class ApiAuthenticationDetailsSource implements AuthenticationDetailsSource<HttpServletRequest, UsernamePasswordAuthenticationToken> {
-
-    private BasicAuthenticationConverter authenticationConverter = new BasicAuthenticationConverter();
-    
-    @Override
-    public UsernamePasswordAuthenticationToken buildDetails(HttpServletRequest arg0) {
-        return authenticationConverter.convert(arg0);
-    }
-
-}
