+++ quadriga/src/main/java/edu/asu/diging/quadriga/config/web/TokenInfo.java
@@ -1,4 +1,4 @@
-package edu.asu.diging.quadriga.api.v1.model;
+package edu.asu.diging.quadriga.config.web;
 
 import java.util.List;
